Finalize task

+++ cplerings-core-infrastructure/src/main/java/com/cplerings/core/infrastructure/repository/CustomOrderRepository.java
@@ -1,8 +1,9 @@
 package com.cplerings.core.infrastructure.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.cplerings.core.domain.order.CustomOrder;
 
-public interface CustomOrderRepository extends JpaRepository<CustomOrder, Integer> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomOrderRepository extends JpaRepository<CustomOrder, Long> {
+
 }
